+++ packages/shared/src/UI/components/index.ts
@@ -2,6 +2,7 @@ export * from './AbstractTab/index.js'
 export * from './AccountIcon/index.js'
 export * from './ActionButton/index.js'
 export * from './AddressItem/index.js'
+export * from './AddCollectibles/index.js'
 export * from './Alert/index.js'
 export * from './ApplicationEntry/index.js'
 export * from './AssetPreviewer/index.js'
